Move BlockingIter into its own file
